@@ -5,7 +5,7 @@ describe('HelloWorld.vue', () => {
   it('renders props.msg when passed', () => {
     const msg = 'new message'
     const wrapper = shallowMount(HelloWorld, {
-      propsData: { msg }
+      props: { msg }
     })
     expect(wrapper.text()).toMatch(msg)
   })
